Fixed duration updating rules
I had to modify how the duration of a given fact is modified. Originally, I had
just modified the fact but with a decremented duration. The problem with this
approach is that the reasserted fact is considered new. So I modified the
duration modification rule to append the word decremented before the rest of
the fact to prevent matching. A new stage has been added
(Advance-Normalization) which removes the symbol decremented from the front of
the fact which restores it to it's former status. However, since this occurs in
a later stage, there is no risk of multiple decrementation actions.