Added a series of messages meant to make cross environment interaction easy
These include run, run-for, eval, build, facts, instances, rules, and assert.

At this point in time there is no make-instance command. That will be
eventually remedied. The difference between run and run-for is that run does
not take any arguments where as run-for takes in a positive number that defines
the number of rules to run the given target for. If you just want to run an
environment then use run because run-for -1 will fail.