Modified SetupEnvironment and CreateExpertSystemRepresentation
The private instances of environments and the QEdict for a given monster are
now given fixed names for simplicity. The QEdict is always known as self and
the environment object is known as env. The public versions of these objects
are still given a gensym for a name within the rhizome but that just makes it
that much easier to figure out what is what within the private environments.