Fix the missing OF symbol
On my gentoo system, my copy of zlib defines _Z_OF instead of OF.
Solving this requires checking to see if OF is defined.
If it is not, then check and see if _Z_OF is also defined. If so, then
define OF in terms of _Z_OF. Fixes compilation errors on my Gentoo amd64
box.